OcDebugAggregatorLib: Renamed to OcLogAggregatorLib
