+++ Docs/Configuration.tex
@@ -3843,7 +3843,7 @@ nvram 4D1FDA02-38C7-4A6A-9CC6-4BCCA8B30102:boot-log |
   \textbf{Libraries}:
   \begin{itemize}
   \tightlist
-  \item \texttt{AAPL} --- OcDebugLogLib, Apple EfiBoot logging
+  \item \texttt{AAPL} --- OcLogAggregatorLib, Apple EfiBoot logging
   \item \texttt{OCABC} --- OcAfterBootCompatLib
   \item \texttt{OCAE} --- OcAppleEventLib
   \item \texttt{OCAK} --- OcAppleKernelLib
@@ -3867,7 +3867,7 @@ nvram 4D1FDA02-38C7-4A6A-9CC6-4BCCA8B30102:boot-log |
   \item \texttt{OCII} --- OcInputLib
   \item \texttt{OCJS} --- OcApfsLib
   \item \texttt{OCKM} --- OcAppleKeyMapLib
-  \item \texttt{OCL} --- OcDebugLogLib
+  \item \texttt{OCL} --- OcLogAggregatorLib
   \item \texttt{OCM} --- OcMiscLib
   \item \texttt{OCMCO} --- OcMachoLib
   \item \texttt{OCME} --- OcHeciLib
