+++ Library/OcAfterBootCompatLib/BootCompatInternal.h
@@ -21,7 +21,6 @@
 
 #include <Library/OcAfterBootCompatLib.h>
 #include <Library/OcBootManagementLib.h>
-#include <Library/OcDebugAggregatorLib.h>
 #include <Library/OcMemoryLib.h>
 
 #include <Protocol/LoadedImage.h>
