+++ Library/OcFlexArrayLib/AsciiStringBuffer.c
@@ -9,7 +9,6 @@
 #include <Library/BaseLib.h>
 #include <Library/MemoryAllocationLib.h>
 #include <Library/DebugLib.h>
-// #include <Library/OcDebugLogLib.h>
 #include <Library/OcFlexArrayLib.h>
 #include <Library/OcGuardLib.h>
 #include <Library/PrintLib.h>
