+++ Library/OcLogAggregatorLib/OcLogAggregatorLib.inf
@@ -18,7 +18,7 @@
 
 [Defines]
   INF_VERSION     = 0x00010005
-  BASE_NAME       = OcDebugAggregatorLib
+  BASE_NAME       = OcLogAggregatorLib
   FILE_GUID       = 2D7559C7-CB17-426A-927D-69BBCD270A9A
   MODULE_TYPE     = BASE
   VERSION_STRING  = 1.0
