+++ Library/OcMainLib/OcMainLib.inf
@@ -81,6 +81,7 @@
   OcFirmwareVolumeLib
   OcGuardLib
   OcHashServicesLib
+  OcLogAggregatorLib
   OcMachoLib
   OcMiscLib
   OcOSInfoLib
