+++ Library/OcMainLib/OpenCoreUefi.c
@@ -36,7 +36,7 @@ WITHOUT WARRANTIES OR REPRESENTATIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED.
 #include <Library/OcConsoleLib.h>
 #include <Library/OcCpuLib.h>
 #include <Library/OcDataHubLib.h>
-#include <Library/OcDebugAggregatorLib.h>
+#include <Library/OcLogAggregatorLib.h>
 #include <Library/OcDebugProtocolLib.h>
 #include <Library/OcDeviceMiscLib.h>
 #include <Library/OcDevicePropertyLib.h>
