+++ Utilities/TestBmf/Makefile
@@ -26,6 +26,7 @@ VPATH   = ../../Platform/OpenCanopy:$\
           ../../Library/OcBootManagementLib:$\
           ../../Library/OcAppleBootPolicyLib:$\
           ../../Library/OcDevicePathLib:$\
+          ../../Library/OcDebugProtocolLib:$\
           ../../Library/OcStringLib:$\
           ../../Library/OcFileLib:$\
           ../../Library/OcStorageLib:$\
